@yuryalencar: Correção de um bug na FrameSalaComercial, e Inicial.
Próxima etapa: Verificar os campos double, para deixar inserir somente
um char == '.'.

+++ JavaApplication19/src/org/unipampa/rp2/frames/FrameSalaComercial.java
@@ -498,11 +498,11 @@ public class FrameSalaComercial extends javax.swing.JFrame {
             }
             
             if(!(jTextFieldValorCondominioSalaComercial.getText().trim().equals(""))){
-                valorCondominio = Integer.parseInt(jTextFieldValorCondominioSalaComercial.getText().trim());
+                valorCondominio = Double.parseDouble(jTextFieldValorCondominioSalaComercial.getText().trim());
             }
             
             if(!(jTextFieldAreaTotalImovel.getText().trim().equals(""))){
-                areaTotal = Integer.parseInt(jTextFieldAreaTotalImovel.getText().trim());
+                areaTotal = Double.parseDouble(jTextFieldAreaTotalImovel.getText().trim());
             }
             
             SalaComercial sala = new SalaComercial(Integer.parseInt(jTextFieldNumeroImovel.getText().trim()),
